+++ native/MuPDFWrapper/MuPDFWrapper.h
@@ -282,7 +282,7 @@ extern "C"
 	/// <param name="out_data">The address of the byte array where the data has been actually written.</param>
 	/// <param name="out_length">The length in bytes of the image data.</param>
 	/// <returns>An integer detailing whether any errors occurred.</returns>
-	DLL_PUBLIC int WriteImage(fz_context* ctx, fz_display_list* list, float x0, float y0, float x1, float y1, float zoom, int colorFormat, int output_format, const fz_buffer** out_buffer, const unsigned char** out_data, size_t* out_length);
+	DLL_PUBLIC int WriteImage(fz_context* ctx, fz_display_list* list, float x0, float y0, float x1, float y1, float zoom, int colorFormat, int output_format, const fz_buffer** out_buffer, const unsigned char** out_data, uint64_t* out_length);
 
 	/// <summary>
 	/// Free a native buffer and its associated resources.
@@ -404,7 +404,7 @@ extern "C"
 	/// <param name="out_image_xres">If the document is an image file, the horizontal resolution of the image.</param>
 	/// <param name="out_image_yres">If the document is an image file, the vertical resolution of the image.</param>
 	/// <returns>An integer detailing whether any errors occurred.</returns>
-	DLL_PUBLIC int CreateDocumentFromStream(fz_context* ctx, const unsigned char* data, const size_t data_length, const char* file_type, int get_image_resolution, const fz_document** out_doc, const fz_stream** out_str, int* out_page_count, float* out_image_xres, float* out_image_yres);
+	DLL_PUBLIC int CreateDocumentFromStream(fz_context* ctx, const unsigned char* data, const uint64_t data_length, const char* file_type, int get_image_resolution, const fz_document** out_doc, const fz_stream** out_str, int* out_page_count, float* out_image_xres, float* out_image_yres);
 
 	/// <summary>
 	/// Free a stream and its associated resources.
@@ -427,14 +427,14 @@ extern "C"
 	/// </summary>
 	/// <param name="ctx">The context whose store's size should be determined.</param>
 	/// <returns>The current size in bytes of the store.</returns>
-	DLL_PUBLIC size_t GetCurrentStoreSize(const fz_context* ctx);
+	DLL_PUBLIC uint64_t GetCurrentStoreSize(const fz_context* ctx);
 
 	/// <summary>
 	/// Get the maximum size of the store.
 	/// </summary>
 	/// <param name="ctx">The context whose store's maximum size should be determined.</param>
 	/// <returns>The maximum size in bytes of the store.</returns>
-	DLL_PUBLIC size_t GetMaxStoreSize(const fz_context* ctx);
+	DLL_PUBLIC uint64_t GetMaxStoreSize(const fz_context* ctx);
 
 	/// <summary>
 	/// Evict items from the store until the total size of the objects in the store is reduced to a given percentage of its current size.
@@ -457,7 +457,7 @@ extern "C"
 	/// <param name="store_size">Maximum size in bytes of the resource store.</param>
 	/// <param name="out_ctx">A pointer to the native context object.</param>
 	/// <returns>An integer detailing whether any errors occurred.</returns>
-	DLL_PUBLIC int CreateContext(long store_size, const fz_context** out_ctx);
+	DLL_PUBLIC int CreateContext(uint64_t store_size, const fz_context** out_ctx);
 
 	/// <summary>
 	/// Free a context and its global store.
